requirements.txt
paho-mqtt
MySQL-python
requests
configparser
